+++ ClubReadUpdater.Lib/SerializeAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace ClubReadUpdater.Lib
+{
+    public class SerializeAttribute : Attribute
+    {
+
+    }
+}
